+++ resources/views/bootstrap-5/includes/bulk-actions.blade.php
@@ -6,7 +6,14 @@
 
         <div class="dropdown-menu dropdown-menu-end w-100" aria-labelledby="bulkActions">
             @foreach($bulkActions as $action => $title)
-                <a href="#" wire:click.prevent="{{ $action }}" class="dropdown-item">{{ $title }}</a>
+                <a
+                    href="#"
+                    wire:click.prevent="{{ $action }}"
+                    wire:key="bulk-action-{{ $action }}"
+                    class="dropdown-item"
+                >
+                    {{ $title }}
+                </a>
             @endforeach
         </div>
     </div>