+++ src/DataTableComponent.php
@@ -136,6 +136,7 @@ abstract class DataTableComponent extends Component
         parent::__construct($id);
 
         $theme = config('livewire-tables.theme');
+
         if ($theme === 'bootstrap-4' || $theme === 'bootstrap-5') {
             $this->paginationTheme = 'bootstrap';
         }